Extracting write back logic from PersonService
Now we have an interface called PersonWriteBackJob and the content of
the method writeBack that was in PersonServiceDefault now resides in
PersonWriteBackJobDefault class, which implements PersonWriteBackJob
interface.

We also have moved the cache key definition from a static value to a
property, so both PersonServiceDefault and PersonWriteBackJobDefault
read the same value from a configuration class name
PersonCacheProperties.

Now that we have a better control on how the cache runs, a new test was
added to verify the behavior of the write back logic is correct without
depending on Thread.sleep method. It uses Awaitibility instead and have
proper assertions.

Unrelated to the refactoring, now the test suite also uses Spring Boot
3.1 feature called ServiceConnections to provision Docker constainers
instead of depending on manual instrumentation and dynamic property. It
declutted the source code considerably.